Moved abstract page class to Anamnesis.Core
